backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
numpy==1.26.3
pillow==10.2.0
onnxruntime==1.24.1
xgboost==2.1.3
scikit-learn==1.4.0
python-jose[cryptography]==3.3.0
